+++ src/main/resources/db/migration/V4__Create_payments_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS  payments (
                           id SERIAL PRIMARY KEY,
-                          order_id INT NOT NULL,
-                          customer_id INT NOT NULL,
+                          order_id BIGINT NOT NULL,
+                          customer_id BIGINT NOT NULL,
                           amount DECIMAL(10,2) NOT NULL,
                           payment_method ENUM('credit_card', 'paypal', 'stripe', 'paystack') NOT NULL,
                           status ENUM('pending', 'completed', 'failed') DEFAULT 'pending',
